fastapi/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.7.1
httpx==0.27.0
